requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
flask_sslify
flask-testing
transformers
torch
scikit-learn
nltk
flask-sqlalchemy
psycopg2-binary
pandas
extract_msg
werkzeug
unstructured[local-inference]
libmagic-dev
poppler-utils
tesseract-ocr
libreoffice
pandocs
paddleocr
pillow
paddleocr
unstructured.paddleocr